+++ torch_frame/datasets/__init__.py
@@ -16,6 +16,7 @@ from .data_frame_benchmark import DataFrameBenchmark
 from .data_frame_text_benchmark import DataFrameTextBenchmark
 from .mercari import Mercari
 from .amazon_fine_food_reviews import AmazonFineFoodReviews
+from .diamond_images import DiamondImages
 from .huggingface_dataset import HuggingFaceDatasetDict
 
 real_world_datasets = [
@@ -34,6 +35,7 @@ real_world_datasets = [
     'DataFrameTextBenchmark',
     'Mercari',
     'AmazonFineFoodReviews',
+    'DiamondImages',
 ]
 
 synthetic_datasets = [